be_acworth_eq_4

Implementation of Acworth et. al. 2016. This is basically the transfer function value for the semi-diurnal band.

be_acworth(
  dat,
  wl = "wl",
  ba = "ba",
  et = "et",
  method = "spec_pgram",
  inverse = TRUE,
  ...
)

Arguments

dat

data that has the independent and dependent variables (data.table)

wl

name of the water level column (character).

ba

name of the barometric pressure column (character).

et

name of the Earth tide column (character).

method

either spec_pgram or spec_welch

inverse

logical whether the barometric relationship is inverse (TRUE means that when the barometric pressure goes up the measured water level goes down (vented transducer, depth to water), FALSE means that when the barometric pressure goes up so does the measured pressure (non-vented transducer)).

...

arguments to pass to transfer_acworth, spec_welch, spec_pgram

Value

barometric efficiency

References

Acworth, R. I., Halloran, L. J. S., Rau, G. C., Cuthbert, M. O., & Bernardi, T. L. (2016). An objective frequency-domain method for quantifying confined aquifer compressible storage using Earth and atmospheric tides. Geophysical Research Letters, 43 (November). doi:10.1002/2016GL071328

Acworth, R. I., G. C. Rau, L. J. S. Halloran, and W. A. Timms (2017), Vertical groundwater stor1077 age properties and changes in confinement determined using hydraulic head response to atmospheric tides, Water Resources Research, 53(4), 2983–2997, doi:10.1002/2016WR020311 .