be_correct

Adjust values based on the barometric efficiency

be_correct(
  dat,
  dep = "wl",
  ind = "baro",
  be = 0,
  inverse = TRUE,
  known_mean = NULL
)

Arguments

dat

data that has the independent and dependent variables (data.table)

dep

name of the dependent variable column (character). This is typically the name for the column holding your water level data.

ind

name of the independent variable column (character). This is typically the name for the column holding your barometric pressure data.

be

numeric value of the barometric efficiency

inverse

whether the barometric relationship is inverse (TRUE means that when the barometric pressure goes up the measured water level goes down (vented transducer, depth to water), FALSE means that when the barometric pressure goes up so does the measured pressure (non-vented transducer)) (logical).

known_mean

numeric explicitly enter the mean if known. Otherwise estimate from data.

Value

numeric vector of corrected values

Examples

library(data.table)
baro = rnorm(1000, sd = 0.01) + 9
wl <- baro * 0.4 + 18
dat <- data.table(baro, wl)
dat$wl + be_correct(dat, be=0.4, inverse = FALSE)
#>    [1] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>    [9] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>   [17] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>   [25] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>   [33] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>   [41] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>   [49] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>   [57] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>   [65] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>   [73] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>   [81] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>   [89] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>   [97] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [105] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [113] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [121] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [129] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [137] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [145] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [153] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [161] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [169] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [177] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [185] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [193] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [201] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [209] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [217] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [225] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [233] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [241] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [249] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [257] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [265] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [273] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [281] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [289] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [297] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [305] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [313] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [321] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [329] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [337] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [345] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [353] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [361] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [369] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [377] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [385] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [393] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [401] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [409] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [417] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [425] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [433] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [441] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [449] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [457] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [465] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [473] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [481] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [489] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [497] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [505] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [513] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [521] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [529] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [537] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [545] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [553] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [561] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [569] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [577] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [585] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [593] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [601] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [609] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [617] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [625] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [633] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [641] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [649] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [657] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [665] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [673] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [681] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [689] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [697] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [705] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [713] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [721] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [729] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [737] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [745] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [753] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [761] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [769] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [777] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [785] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [793] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [801] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [809] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [817] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [825] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [833] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [841] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [849] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [857] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [865] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [873] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [881] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [889] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [897] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [905] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [913] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [921] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [929] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [937] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [945] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [953] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [961] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [969] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [977] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [985] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
#>  [993] 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994 21.59994
dat$wl + be_correct(dat, be=0.4, inverse = FALSE, known_mean = 9) 
#>    [1] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>   [15] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>   [29] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>   [43] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>   [57] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>   [71] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>   [85] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>   [99] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [113] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [127] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [141] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [155] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [169] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [183] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [197] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [211] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [225] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [239] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [253] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [267] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [281] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [295] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [309] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [323] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [337] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [351] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [365] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [379] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [393] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [407] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [421] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [435] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [449] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [463] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [477] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [491] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [505] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [519] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [533] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [547] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [561] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [575] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [589] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [603] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [617] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [631] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [645] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [659] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [673] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [687] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [701] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [715] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [729] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [743] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [757] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [771] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [785] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [799] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [813] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [827] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [841] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [855] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [869] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [883] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [897] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [911] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [925] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [939] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [953] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [967] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [981] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [995] 21.6 21.6 21.6 21.6 21.6 21.6
# should return ~21.6 = 18 + 0.4 * 9