be_visual

Generate dataset for comparing barometric efficiency

be_visual(
  dat,
  dep = "wl",
  ind = "baro",
  time = "datetime",
  be_tests = seq(0, 1, 0.1),
  inverse = TRUE,
  subsample = TRUE
)

Arguments

dat

data that has the independent and dependent variables (data.table)

dep

name of the dependent variable column (character). This is typically the name for the column holding your water level data.

ind

name of the independent variable column (character). This is typically the name for the column holding your barometric pressure data.

time

name of the column containing the time (character)

be_tests

vector of barometric efficiencies to test (between 0 and 1) (numeric)

inverse

whether the barometric relationship is inverse (TRUE means that when the barometric pressure goes up the measured water level goes down (vented transducer, depth to water), FALSE means that when the barometric pressure goes up so does the measured pressure (non-vented transducer)) (logical).

subsample

should the data be subsampled for plotting? (logical)

Value

data.table of barometric efficiency compensated datasets

References

Smith, L. A., van der Kamp, G., & Hendry, M. J. (2013). A new technique for obtaining high‐resolution pore pressure records in thick claystone aquitards and its use to determine in situ compressibility. Water Resources Research, 49(2), 732-743. doi:10.1002/wrcr.20084

Examples

library(data.table)
be <- 0.43
x <- seq(0, 28*pi, pi / (12*12))

baro <- sin(x) + rnorm(length(x), sd = 0.04)
wl <- -sin(x) * be + rnorm(length(x), sd = 0.04)
dat <- data.table(datetime = as.POSIXct(x * 86400 / (2 * pi),
                                        origin = '1970-01-01', tz = 'UTC'),
                  wl = wl, baro = baro)
be_visual(dat)