lag_difference

Calculate differences for the dependent and independent variables and remove any NA values. This modifies the data.table in place.

lag_difference(
  dat,
  var_name = "wl",
  lag_space = 1,
  inverse = FALSE,
  remove_na = TRUE
)

Arguments

dat

data that has the independent and dependent variables (data.table)

var_name

name of the column to lag (character)

lag_space

space between difference calculation in number of observations

inverse

whether the barometric relationship is inverse (TRUE means that when the barometric pressure goes up the measured water level goes down (vented transducer, depth to water), FALSE means that when the barometric pressure goes up so does the measured pressure (non-vented transducer)) (logical).

remove_na

remove NA values (logical)

Value

data.table with lagged differences

Examples

library(data.table)
datetime <- seq.POSIXt(as.POSIXct("2016-01-01 12:00:00"),
                       as.POSIXct("2016-01-05 12:00:00"), by='hour' )
baro <- sin(seq(0, 2*pi, length.out = length(datetime)))
wl <- 0.4 * baro
dat <- data.table(baro, wl, datetime)
lag_difference(dat, lag_space = 1, remove_na = FALSE)
#>              baro            wl            datetime
#>             <num>         <num>              <POSc>
#>  1:  0.000000e+00            NA 2016-01-01 12:00:00
#>  2:  6.540313e-02  0.0261612517 2016-01-01 13:00:00
#>  3:  1.305262e-01  0.0260492252 2016-01-01 14:00:00
#>  4:  1.950903e-01  0.0258256519 2016-01-01 15:00:00
#>  5:  2.588190e-01  0.0254914892 2016-01-01 16:00:00
#>  6:  3.214395e-01  0.0250481681 2016-01-01 17:00:00
#>  7:  3.826834e-01  0.0244975868 2016-01-01 18:00:00
#>  8:  4.422887e-01  0.0238421031 2016-01-01 19:00:00
#>  9:  5.000000e-01  0.0230845239 2016-01-01 20:00:00
#> 10:  5.555702e-01  0.0222280932 2016-01-01 21:00:00
#> 11:  6.087614e-01  0.0212764784 2016-01-01 22:00:00
#> 12:  6.593458e-01  0.0202337544 2016-01-01 23:00:00
#> 13:  7.071068e-01  0.0191043864 2016-01-02 00:00:00
#> 14:  7.518398e-01  0.0178932105 2016-01-02 01:00:00
#> 15:  7.933533e-01  0.0166054131 2016-01-02 02:00:00
#> 16:  8.314696e-01  0.0152465088 2016-01-02 03:00:00
#> 17:  8.660254e-01  0.0138223166 2016-01-02 04:00:00
#> 18:  8.968727e-01  0.0123389351 2016-01-02 05:00:00
#> 19:  9.238795e-01  0.0108027164 2016-01-02 06:00:00
#> 20:  9.469301e-01  0.0092202388 2016-01-02 07:00:00
#> 21:  9.659258e-01  0.0075982787 2016-01-02 08:00:00
#> 22:  9.807853e-01  0.0059437816 2016-01-02 09:00:00
#> 23:  9.914449e-01  0.0042638324 2016-01-02 10:00:00
#> 24:  9.978589e-01  0.0025656247 2016-01-02 11:00:00
#> 25:  1.000000e+00  0.0008564307 2016-01-02 12:00:00
#> 26:  9.978589e-01 -0.0008564307 2016-01-02 13:00:00
#> 27:  9.914449e-01 -0.0025656247 2016-01-02 14:00:00
#> 28:  9.807853e-01 -0.0042638324 2016-01-02 15:00:00
#> 29:  9.659258e-01 -0.0059437816 2016-01-02 16:00:00
#> 30:  9.469301e-01 -0.0075982787 2016-01-02 17:00:00
#> 31:  9.238795e-01 -0.0092202388 2016-01-02 18:00:00
#> 32:  8.968727e-01 -0.0108027164 2016-01-02 19:00:00
#> 33:  8.660254e-01 -0.0123389351 2016-01-02 20:00:00
#> 34:  8.314696e-01 -0.0138223166 2016-01-02 21:00:00
#> 35:  7.933533e-01 -0.0152465088 2016-01-02 22:00:00
#> 36:  7.518398e-01 -0.0166054131 2016-01-02 23:00:00
#> 37:  7.071068e-01 -0.0178932105 2016-01-03 00:00:00
#> 38:  6.593458e-01 -0.0191043864 2016-01-03 01:00:00
#> 39:  6.087614e-01 -0.0202337544 2016-01-03 02:00:00
#> 40:  5.555702e-01 -0.0212764784 2016-01-03 03:00:00
#> 41:  5.000000e-01 -0.0222280932 2016-01-03 04:00:00
#> 42:  4.422887e-01 -0.0230845239 2016-01-03 05:00:00
#> 43:  3.826834e-01 -0.0238421031 2016-01-03 06:00:00
#> 44:  3.214395e-01 -0.0244975868 2016-01-03 07:00:00
#> 45:  2.588190e-01 -0.0250481681 2016-01-03 08:00:00
#> 46:  1.950903e-01 -0.0254914892 2016-01-03 09:00:00
#> 47:  1.305262e-01 -0.0258256519 2016-01-03 10:00:00
#> 48:  6.540313e-02 -0.0260492252 2016-01-03 11:00:00
#> 49:  1.224647e-16 -0.0261612517 2016-01-03 12:00:00
#> 50: -6.540313e-02 -0.0261612517 2016-01-03 13:00:00
#> 51: -1.305262e-01 -0.0260492252 2016-01-03 14:00:00
#> 52: -1.950903e-01 -0.0258256519 2016-01-03 15:00:00
#> 53: -2.588190e-01 -0.0254914892 2016-01-03 16:00:00
#> 54: -3.214395e-01 -0.0250481681 2016-01-03 17:00:00
#> 55: -3.826834e-01 -0.0244975868 2016-01-03 18:00:00
#> 56: -4.422887e-01 -0.0238421031 2016-01-03 19:00:00
#> 57: -5.000000e-01 -0.0230845239 2016-01-03 20:00:00
#> 58: -5.555702e-01 -0.0222280932 2016-01-03 21:00:00
#> 59: -6.087614e-01 -0.0212764784 2016-01-03 22:00:00
#> 60: -6.593458e-01 -0.0202337544 2016-01-03 23:00:00
#> 61: -7.071068e-01 -0.0191043864 2016-01-04 00:00:00
#> 62: -7.518398e-01 -0.0178932105 2016-01-04 01:00:00
#> 63: -7.933533e-01 -0.0166054131 2016-01-04 02:00:00
#> 64: -8.314696e-01 -0.0152465088 2016-01-04 03:00:00
#> 65: -8.660254e-01 -0.0138223166 2016-01-04 04:00:00
#> 66: -8.968727e-01 -0.0123389351 2016-01-04 05:00:00
#> 67: -9.238795e-01 -0.0108027164 2016-01-04 06:00:00
#> 68: -9.469301e-01 -0.0092202388 2016-01-04 07:00:00
#> 69: -9.659258e-01 -0.0075982787 2016-01-04 08:00:00
#> 70: -9.807853e-01 -0.0059437816 2016-01-04 09:00:00
#> 71: -9.914449e-01 -0.0042638324 2016-01-04 10:00:00
#> 72: -9.978589e-01 -0.0025656247 2016-01-04 11:00:00
#> 73: -1.000000e+00 -0.0008564307 2016-01-04 12:00:00
#> 74: -9.978589e-01  0.0008564307 2016-01-04 13:00:00
#> 75: -9.914449e-01  0.0025656247 2016-01-04 14:00:00
#> 76: -9.807853e-01  0.0042638324 2016-01-04 15:00:00
#> 77: -9.659258e-01  0.0059437816 2016-01-04 16:00:00
#> 78: -9.469301e-01  0.0075982787 2016-01-04 17:00:00
#> 79: -9.238795e-01  0.0092202388 2016-01-04 18:00:00
#> 80: -8.968727e-01  0.0108027164 2016-01-04 19:00:00
#> 81: -8.660254e-01  0.0123389351 2016-01-04 20:00:00
#> 82: -8.314696e-01  0.0138223166 2016-01-04 21:00:00
#> 83: -7.933533e-01  0.0152465088 2016-01-04 22:00:00
#> 84: -7.518398e-01  0.0166054131 2016-01-04 23:00:00
#> 85: -7.071068e-01  0.0178932105 2016-01-05 00:00:00
#> 86: -6.593458e-01  0.0191043864 2016-01-05 01:00:00
#> 87: -6.087614e-01  0.0202337544 2016-01-05 02:00:00
#> 88: -5.555702e-01  0.0212764784 2016-01-05 03:00:00
#> 89: -5.000000e-01  0.0222280932 2016-01-05 04:00:00
#> 90: -4.422887e-01  0.0230845239 2016-01-05 05:00:00
#> 91: -3.826834e-01  0.0238421031 2016-01-05 06:00:00
#> 92: -3.214395e-01  0.0244975868 2016-01-05 07:00:00
#> 93: -2.588190e-01  0.0250481681 2016-01-05 08:00:00
#> 94: -1.950903e-01  0.0254914892 2016-01-05 09:00:00
#> 95: -1.305262e-01  0.0258256519 2016-01-05 10:00:00
#> 96: -6.540313e-02  0.0260492252 2016-01-05 11:00:00
#> 97: -2.449294e-16  0.0261612517 2016-01-05 12:00:00
#>              baro            wl            datetime