be_correct

Adjust values based on the barometric efficiency

be_correct(dat, dep = "wl", ind = "baro", be = 0, inverse = TRUE,
  known_mean = NULL)

Arguments

dat
data that has the independent and dependent variables (data.table)

dep
name of the dependent variable column (character). This is typically the name for the column holding your water level data.

ind
name of the independent variable column (character). This is typically the name for the column holding your barometric pressure data.

be
numeric value of the barometric efficiency

inverse
whether the barometric relationship is inverse (TRUE means that when the barometric pressure goes up the measured water level goes down (vented transducer, depth to water), FALSE means that when the barometric pressure goes up so does the measured pressure (non-vented transducer)) (logical).

known_mean
numeric explicitly enter the mean if known. Otherwise estimate from data.

Value

numeric vector of corrected values

Examples

library(data.table)
baro = rnorm(1000, sd = 0.01) + 9
wl <- baro * 0.4 + 18
dat <- data.table(baro, wl)
dat$wl + be_correct(dat, be=0.4, inverse = FALSE)

#>    [1] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>    [9] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>   [17] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>   [25] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>   [33] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>   [41] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>   [49] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>   [57] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>   [65] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>   [73] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>   [81] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>   [89] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>   [97] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [105] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [113] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [121] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [129] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [137] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [145] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [153] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [161] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [169] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [177] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [185] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [193] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [201] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [209] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [217] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [225] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [233] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [241] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [249] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [257] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [265] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [273] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [281] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [289] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [297] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [305] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [313] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [321] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [329] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [337] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [345] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [353] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [361] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [369] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [377] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [385] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [393] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [401] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [409] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [417] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [425] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [433] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [441] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [449] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [457] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [465] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [473] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [481] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [489] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [497] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [505] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [513] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [521] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [529] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [537] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [545] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [553] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [561] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [569] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [577] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [585] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [593] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [601] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [609] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [617] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [625] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [633] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [641] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [649] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [657] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [665] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [673] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [681] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [689] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [697] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [705] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [713] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [721] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [729] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [737] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [745] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [753] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [761] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [769] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [777] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [785] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [793] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [801] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [809] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [817] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [825] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [833] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [841] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [849] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [857] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [865] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [873] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [881] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [889] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [897] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [905] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [913] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [921] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [929] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [937] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [945] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [953] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [961] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [969] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [977] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [985] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993
#>  [993] 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993 21.59993

dat$wl + be_correct(dat, be=0.4, inverse = FALSE, known_mean = 9)

#>    [1] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>   [15] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>   [29] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>   [43] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>   [57] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>   [71] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>   [85] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>   [99] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [113] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [127] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [141] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [155] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [169] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [183] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [197] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [211] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [225] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [239] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [253] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [267] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [281] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [295] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [309] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [323] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [337] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [351] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [365] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [379] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [393] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [407] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [421] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [435] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [449] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [463] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [477] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [491] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [505] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [519] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [533] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [547] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [561] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [575] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [589] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [603] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [617] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [631] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [645] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [659] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [673] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [687] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [701] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [715] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [729] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [743] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [757] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [771] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [785] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [799] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [813] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [827] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [841] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [855] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [869] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [883] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [897] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [911] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [925] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [939] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [953] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [967] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [981] 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6 21.6
#>  [995] 21.6 21.6 21.6 21.6 21.6 21.6

# should return ~21.6 = 18 + 0.4 * 9